requirements.txt
Babel==2.5.1
Django==1.11
django-common-helpers==0.9.1
django-cors-headers==2.1.0
django-cron==0.5.0
django-crontab==0.7.1
django-enumfield==1.3b2
django-filter==1.0.4
django-invitations==1.9.1
django-modeltranslation==0.12.1
django-phonenumber-field==1.3.0
django-rest-framework-docs==0.1.7
django-sendsms==0.2.3
django-twilio==0.8.0
djangorestframework==3.6.4
drfdocs==0.0.11
geopy==1.11.0
jsonfield==2.0.2
jsonpickle==0.9.5
Markdown==2.6.9
olefile==0.44
phonenumberslite==8.8.8
Pillow==4.2.1
psycopg2==2.7.3.1
PyJWT==1.5.3
PySocks==1.6.7
python-dateutil==2.5.3
pytz==2017.2
requests==2.11.1
#Shapely==1.6.1
six==1.11.0
twilio==6.10.0
XlsxWriter==1.0.0
djutils==0.3.2
django-unused-media==0.1.10
urllib3==1.22
xlrd==1.1.0
reportlab==3.4.0
inflect==0.3.1
